# app/models/JobApplicationState.py
# coding: utf-8

from __future__ import absolute_import

from datetime import date, datetime  # noqa: F401
from typing import List, Dict  # noqa: F401

from models.BaseModel import BaseModel


class JobApplicationState(BaseModel):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    """
    allowed enum values
    """
    NEW = "NEW"
    ACTIVE = "ACTIVE"
    REJECTED = "REJECTED"
    ACCEPTED = "ACCEPTED"
    WITHDRAWN = "WITHDRAWN"

    @staticmethod
    def is_readonly(state):
        return state == JobApplicationState.REJECTED or \
               state == JobApplicationState.ACCEPTED or \
               state == JobApplicationState.WITHDRAWN