# app/models/UpdateJobApplicationDto.py
# coding: utf-8

from __future__ import absolute_import

from datetime import date, datetime  # noqa: F401
from typing import List, Dict  # noqa: F401

from models.BaseModel import BaseModel

note = 'note'


class UpdateJobApplicationDto(BaseModel):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, new_note: str = None):  # noqa: E501
        """UpdateJobApplicationDto - a model defined in Swagger

        :param new_note: The note of this UpdateJobApplicationDto.  # noqa: E501
        :type new_note: str
        """
        self._data = {
            note: new_note
        }

    @classmethod
    def from_dict(cls, dikt) -> 'UpdateJobApplicationDto':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UpdateJobApplicationDto of this UpdateJobApplicationDto.  # noqa: E501
        :rtype: UpdateJobApplicationDto
        """
        instance = cls()
        instance.note = dikt[note]
        return instance

    @property
    def note(self) -> str:
        """Gets the note of this UpdateJobApplicationDto.


        :return: The note of this UpdateJobApplicationDto.
        :rtype: str
        """
        return self._data[note]

    @note.setter
    def note(self, new_note: str):
        """Sets the note of this UpdateJobApplicationDto.


        :param new_note: The note of this UpdateJobApplicationDto.
        :type new_note: str
        """
        self._data[note] = new_note